+++ environment/core/hlt.hpp
@@ -29,7 +29,7 @@ namespace hlt {
     struct Velocity {
         short vel_x, vel_y;
 
-        auto accelerate_by(unsigned short magnitude, unsigned short angle) -> void;
+        auto accelerate_by(unsigned short magnitude, double angle) -> void;
     };
 
     static bool operator==(const Location& l1, const Location& l2) {
